Rename stats lib file -> memberStats
