+++ src/commands/pill.ts
@@ -4,7 +4,7 @@ import { botColor } from "../lib/util";
 import { redis } from "../lib/redis";
 import { pills } from "../lib/data/pills";
 import { MemberStats } from "../lib/data/stats";
-import { GetMemberStat, GetMemberStatsEmbed } from "../lib/stats";
+import { GetMemberStat, GetMemberStatsEmbed } from "../lib/memberStats";
 import * as dotenv from "dotenv";
 dotenv.config();
